chore: extract `AssertionResult` from `@jest/test-result`
