+++ packages/jest-message-util/package.json
@@ -21,7 +21,6 @@
   },
   "dependencies": {
     "@babel/code-frame": "^7.0.0",
-    "@jest/test-result": "^25.2.4",
     "@jest/types": "^25.2.3",
     "@types/stack-utils": "^1.0.1",
     "chalk": "^3.0.0",
