+++ packages/jest-test-result/src/types.ts
@@ -7,7 +7,7 @@
 
 import type {CoverageMap, CoverageMapData} from 'istanbul-lib-coverage';
 import type {ConsoleBuffer} from '@jest/console';
-import type {Config, TransformTypes} from '@jest/types';
+import type {Config, TestResult, TransformTypes} from '@jest/types';
 import type {V8Coverage} from 'collect-v8-coverage';
 
 export type V8CoverageResult = Array<{
@@ -15,12 +15,7 @@ export type V8CoverageResult = Array<{
   result: V8Coverage[number];
 }>;
 
-export type SerializableError = {
-  code?: unknown;
-  message: string;
-  stack: string | null | undefined;
-  type?: string;
-};
+export type SerializableError = TestResult.SerializableError;
 
 export type FailedAssertion = {
   matcherName?: string;
@@ -39,41 +34,19 @@ export type AssertionLocation = {
   path: string;
 };
 
-export type Status =
-  | 'passed'
-  | 'failed'
-  | 'skipped'
-  | 'pending'
-  | 'todo'
-  | 'disabled';
+export type Status = TestResult.Status;
 
 export type Bytes = number;
 
-export type Milliseconds = number;
-type Callsite = {
-  column: number;
-  line: number;
-};
+export type Milliseconds = TestResult.Milliseconds;
 
-export type AssertionResult = {
-  ancestorTitles: Array<string>;
-  duration?: Milliseconds | null | undefined;
-  failureMessages: Array<string>;
-  fullName: string;
-  invocations?: number;
-  location: Callsite | null | undefined;
-  numPassingAsserts: number;
-  status: Status;
-  title: string;
-};
+export type AssertionResult = TestResult.AssertionResult;
 
-export type FormattedAssertionResult = {
-  ancestorTitles: Array<string>;
-  failureMessages: Array<string> | null;
-  fullName: string;
-  location: Callsite | null | undefined;
-  status: Status;
-  title: string;
+export type FormattedAssertionResult = Pick<
+  AssertionResult,
+  'ancestorTitles' | 'fullName' | 'location' | 'status' | 'title'
+> & {
+  failureMessages: AssertionResult['failureMessages'] | null;
 };
 
 export type AggregatedResultWithoutCoverage = {
@@ -135,7 +108,7 @@ export type TestResult = {
     [sourcePath: string]: string;
   };
   testExecError?: SerializableError;
-  testFilePath: string;
+  testFilePath: Config.Path;
   testResults: Array<AssertionResult>;
   v8Coverage?: V8CoverageResult;
 };
