+++ packages/jest-types/src/index.ts
@@ -8,6 +8,7 @@
 import type * as Circus from './Circus';
 import type * as Config from './Config';
 import type * as Global from './Global';
+import type * as TestResult from './TestResult';
 import type * as TransformTypes from './Transform';
 
-export type {Circus, Config, Global, TransformTypes};
+export type {Circus, Config, Global, TestResult, TransformTypes};
